requirements.txt
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
python-dotenv==1.0.0
requests==2.30.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.6
